MAINT: refactored scaling methods of MCO, moving the MCO-specific function to a class method. Now `kpi_scaling.py` should only provide generic scaling methods.
